Moved Pageview to its own file under modals, moved calendar utils into the utils folder
