+++ lib/pages/home.dart
@@ -5,6 +5,7 @@ import 'dart:ui';
 
 import 'package:date_format/date_format.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_launcher/pages/modals/pageview.dart';
 import 'package:flutter_launcher/pages/settings.dart';
 import 'package:flutter_launcher/widgets/utils/widget_utils.dart';
 import 'package:flutter_launcher/widgets/utils/widgetchangenotifier.dart';
@@ -123,7 +124,6 @@ class _launcherState extends State<launcher>{
     String? appName2 = widget.prefs.getString("Pinned App2");
     String? appName3 = widget.prefs.getString("Pinned App3");
     String? appName4 = widget.prefs.getString("Pinned App4");
-    int? restoreLastPage = widget.prefs.getInt("Page");
     
     if (togglePinApp != null){
       pinAppToggle(togglePinApp);
@@ -183,16 +183,6 @@ class _launcherState extends State<launcher>{
         searchHieght = 40;
       });     
     }
-    if (restoreLastPage != null){
-      setState(() {
-        lastPage = restoreLastPage;
-      });
-    }
-    WidgetList widgets = WidgetList(widgets: [], prefs: widget.prefs);
-    await widgets.loadWidgets();
-    setState(() {
-     initialItems = widgets.widgets;
-    });
   }
 
   void fetchApps() async {
@@ -368,46 +358,7 @@ class _launcherState extends State<launcher>{
                     onVerticalDragStart: (details) {
                       showModalBottomSheet<void>(isScrollControlled: true ,showDragHandle: true ,context: context, builder: (BuildContext context) {
                         return StatefulBuilder(builder: (BuildContext context, StateSetter setState) {
-                          final visibilityState = Provider.of<WidgetVisibilityState>(context);
-
-                          Future<List<Widget>> getVisibleWidgets() async {
-                            return visibilityState.order
-                           .where((index) => visibilityState.visibility[index])
-                           .map((index) => initialItems[index])
-                           .toList();
-                          }
-
-                          return FutureBuilder<List<Widget>>(
-                            future: getVisibleWidgets(),
-                            builder: (context, snapshot) {
-                              if (snapshot.connectionState == ConnectionState.waiting) {
-                                return Widgetoptions(widget.prefs);
-                              } else if (snapshot.hasError) {
-                                return Text('Error: ${snapshot.error}');
-                              } else {
-                                List<Widget> visibleWidgets = snapshot.data ?? [];
-                                PageController _pageController = PageController(initialPage: lastPage);
-
-                                return Padding(
-                                  padding: EdgeInsets.only(bottom: MediaQuery.of(context).viewInsets.bottom),
-                                  child: SizedBox(
-                                    height: 500,
-                                    child: PageView(
-                                      controller: _pageController,
-                                      onPageChanged: (int page){
-                                        lastPage = page;
-                                        widget.prefs.setInt("Page", page);
-                                      },
-                                      children: [
-                                        Widgetoptions(widget.prefs),
-                                        ...visibleWidgets,
-                                      ],
-                                    ),
-                                 ),
-                                );
-                              }
-                            }
-                          );
+                          return pages(widget.prefs);
                         });
                       });
                     },
